requirements.txt
# Flask core
flask>=3.1.0
flask-bcrypt>=1.0.1
flask-jwt-extended>=4.7.1
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.4
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.6
SQLAlchemy>=2.0.0
marshmallow>=3.19.0

# Server / Deployment
gunicorn>=23.0.0

# Testing
pytest>=7.4.0
pytest-flask>=1.2.0

# Utils
click>=8.1.8
colorama>=0.4.6
blinker>=1.9.0
itsdangerous>=2.2.0
jinja2>=3.1.6
markupsafe>=3.0.2
greenlet>=3.1.1
packaging>=24.2
pyjwt>=2.10.1
setuptools>=65.5.0
typing-extensions>=4.12.2
pip>=24.0
alembic==1.11.3
